requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-allauth==65.1.0
django-crispy-forms==2.3
gTTS==2.5.4
gunicorn==23.0.0
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3